Improvements on notification
