+++ go.mod
@@ -2,8 +2,6 @@ module github.com/PJNube/lib-nats
 
 go 1.24.0
 
-toolchain go1.24.5
-
 require (
 	github.com/nats-io/nats.go v1.39.1
 	github.com/rs/xid v1.6.0
@@ -11,14 +9,6 @@ require (
 )
 
 require (
-	github.com/jinzhu/inflection v1.0.0 // indirect
-	github.com/jinzhu/now v1.1.5 // indirect
-	golang.org/x/text v0.21.0 // indirect
-	gorm.io/gorm v1.25.12 // indirect
-)
-
-require (
-	github.com/PJNube/lib-models v0.0.14
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/klauspost/compress v1.17.9 // indirect
 	github.com/nats-io/nkeys v0.4.9 // indirect
