+++ pjnnats/base.go
@@ -3,7 +3,6 @@ package pjnnats
 import (
 	"context"
 	"crypto/tls"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
@@ -13,7 +12,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/PJNube/lib-models/dtos"
 	"github.com/nats-io/nats.go"
 	"github.com/rs/xid"
 )
@@ -23,7 +21,6 @@ type ConnectionStatus int
 const (
 	Active ConnectionStatus = iota
 	Disabled
-	NatsConnectionUUID = "local"
 )
 
 // Opts represents options for NATS operations.
@@ -381,32 +378,6 @@ func (n *Client) Publish(uuid string, msg nats.Msg) error {
 		return err
 	}
 	return conn.Connection.PublishMsg(&msg)
-}
-
-func (n *Client) PublishNotification(subject, event string, message any) error {
-	payload, err := getNotifyPayload(message, event)
-	if err != nil {
-		return err
-	}
-
-	return n.Publish(NatsConnectionUUID, nats.Msg{
-		Subject: subject,
-		Data:    payload,
-	})
-}
-
-func getNotifyPayload(message any, event string) ([]byte, error) {
-	notifyPayload := &dtos.NotificationPayload{
-		Event: event,
-		Data:  message,
-	}
-
-	payload, err := json.Marshal(notifyPayload)
-	if err != nil {
-		return nil, fmt.Errorf("error marshalling notification payload: %w", err)
-	}
-
-	return payload, nil
 }
 
 // Subscribe listens for messages on a specific connection.
